+++ internal/scm/gitea/gitea.go
@@ -36,6 +36,8 @@ func init() {
 		RequiresProject:       true,
 		RequiresAPIKey:        true,
 		ValidateTrackerConfig: validateConfig,
+		DefaultActiveStates:   defaultActiveStates,
+		DefaultTerminalStates: defaultTerminalStates,
 	})
 }